# backend/docs/openapi_schema.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TrainerPlatform",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/certification-core/blueprints": {
      "get": {
        "tags": [
          "Certification-Core",
          "Certification-Core"
        ],
        "summary": "List Blueprints",
        "description": "List exam blueprints with optional filters.",
        "operationId": "list_blueprints_api_v1_certification_core_blueprints_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Status"
            }
          },
          {
            "name": "domain_pack_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Domain Pack Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExamBlueprintListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Certification-Core",
          "Certification-Core"
        ],
        "summary": "Create Blueprint",
        "description": "Create a new exam blueprint with sections.",
        "operationId": "create_blueprint_api_v1_certification_core_blueprints_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExamBlueprintCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExamBlueprintResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/certification-core/blueprints/{blueprint_id}": {
      "get": {
        "tags": [
          "Certification-Core",
          "Certification-Core"
        ],
        "summary": "Get Blueprint",
        "description": "Get an exam blueprint by its business ID.",
        "operationId": "get_blueprint_api_v1_certification_core_blueprints__blueprint_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "blueprint_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Blueprint Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExamBlueprintResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Certification-Core",
          "Certification-Core"
        ],
        "summary": "Update Blueprint",
        "description": "Update an exam blueprint.",
        "operationId": "update_blueprint_api_v1_certification_core_blueprints__blueprint_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "blueprint_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Blueprint Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExamBlueprintUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExamBlueprintResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/certification-core/audit": {
      "get": {
        "tags": [
          "Certification-Core",
          "Certification-Core"
        ],
        "summary": "Query Audit",
        "description": "Query audit events with filters. Requires audit:read permission.",
        "operationId": "query_audit_api_v1_certification_core_audit_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "entity_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Entity Type"
            }
          },
          {
            "name": "entity_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Entity Id"
            }
          },
          {
            "name": "actor_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Actor Id"
            }
          },
          {
            "name": "action",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Action"
            }
          },
          {
            "name": "date_from",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date From"
            }
          },
          {
            "name": "date_to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date To"
            }
          },
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuditEventListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AuditEventListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AuditEventResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "skip": {
            "type": "integer",
            "title": "Skip",
            "default": 0
          },
          "limit": {
            "type": "integer",
            "title": "Limit",
            "default": 100
          }
        },
        "type": "object",
        "required": [
          "items",
          "total"
        ],
        "title": "AuditEventListResponse"
      },
      "AuditEventResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "audit_event_id": {
            "type": "string",
            "title": "Audit Event Id"
          },
          "entity_type": {
            "type": "string",
            "title": "Entity Type"
          },
          "entity_id": {
            "type": "string",
            "title": "Entity Id"
          },
          "entity_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Entity Version"
          },
          "action": {
            "type": "string",
            "title": "Action"
          },
          "actor_id": {
            "type": "string",
            "title": "Actor Id"
          },
          "actor_role": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Actor Role"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "before_hash": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Before Hash"
          },
          "after_hash": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "After Hash"
          },
          "event_timestamp": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Event Timestamp"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "audit_event_id",
          "entity_type",
          "entity_id",
          "action",
          "actor_id"
        ],
        "title": "AuditEventResponse"
      },
      "BlueprintSectionCreate": {
        "properties": {
          "section_id": {
            "type": "string",
            "maxLength": 100,
            "title": "Section Id"
          },
          "name": {
            "type": "string",
            "maxLength": 255,
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "competency_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Competency Ids",
            "default": []
          },
          "weight_percent": {
            "type": "number",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Weight Percent"
          },
          "minimum_items": {
            "type": "integer",
            "title": "Minimum Items",
            "default": 0
          },
          "maximum_items": {
            "type": "integer",
            "title": "Maximum Items",
            "default": 0
          },
          "difficulty_distribution": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Difficulty Distribution"
          },
          "cognitive_distribution": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cognitive Distribution"
          },
          "critical_section": {
            "type": "boolean",
            "title": "Critical Section",
            "default": false
          },
          "sort_order": {
            "type": "integer",
            "title": "Sort Order",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "section_id",
          "name",
          "weight_percent"
        ],
        "title": "BlueprintSectionCreate"
      },
      "BlueprintSectionResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "blueprint_id": {
            "type": "string",
            "title": "Blueprint Id"
          },
          "section_id": {
            "type": "string",
            "title": "Section Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "competency_ids": {
            "anyOf": [
              {
                "items": {},
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Competency Ids"
          },
          "weight_percent": {
            "type": "number",
            "title": "Weight Percent"
          },
          "minimum_items": {
            "type": "integer",
            "title": "Minimum Items"
          },
          "maximum_items": {
            "type": "integer",
            "title": "Maximum Items"
          },
          "difficulty_distribution": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Difficulty Distribution"
          },
          "cognitive_distribution": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cognitive Distribution"
          },
          "critical_section": {
            "type": "boolean",
            "title": "Critical Section"
          },
          "sort_order": {
            "type": "integer",
            "title": "Sort Order"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "blueprint_id",
          "section_id",
          "name",
          "weight_percent",
          "minimum_items",
          "maximum_items",
          "critical_section",
          "sort_order"
        ],
        "title": "BlueprintSectionResponse"
      },
...
      "ExamBlueprintCreate": {
        "properties": {
          "blueprint_id": {
            "type": "string",
            "maxLength": 100,
            "title": "Blueprint Id"
          },
          "domain_pack_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Domain Pack Id"
          },
          "competency_framework_version": {
            "type": "string",
            "maxLength": 100,
            "title": "Competency Framework Version"
          },
          "version": {
            "type": "string",
            "maxLength": 20,
            "title": "Version"
          },
          "exam_duration_minutes": {
            "type": "integer",
            "title": "Exam Duration Minutes",
            "default": 60
          },
          "total_items": {
            "type": "integer",
            "title": "Total Items",
            "default": 0
          },
          "pass_policy_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pass Policy Id"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "sections": {
            "items": {
              "$ref": "#/components/schemas/BlueprintSectionCreate"
            },
            "type": "array",
            "title": "Sections",
            "default": []
          },
          "created_by": {
            "type": "string",
            "maxLength": 100,
            "title": "Created By"
          }
        },
        "type": "object",
        "required": [
          "blueprint_id",
          "competency_framework_version",
          "version",
          "created_by"
        ],
        "title": "ExamBlueprintCreate"
      },
      "ExamBlueprintListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ExamBlueprintResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "skip": {
            "type": "integer",
            "title": "Skip",
            "default": 0
          },
          "limit": {
            "type": "integer",
            "title": "Limit",
            "default": 100
          }
        },
        "type": "object",
        "required": [
          "items",
          "total"
        ],
        "title": "ExamBlueprintListResponse"
      },
      "ExamBlueprintResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "blueprint_id": {
            "type": "string",
            "title": "Blueprint Id"
          },
          "domain_pack_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Domain Pack Id"
          },
          "competency_framework_version": {
            "type": "string",
            "title": "Competency Framework Version"
          },
          "version": {
            "type": "string",
            "title": "Version"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "exam_duration_minutes": {
            "type": "integer",
            "title": "Exam Duration Minutes"
          },
          "total_items": {
            "type": "integer",
            "title": "Total Items"
          },
          "pass_policy_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pass Policy Id"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "created_by": {
            "type": "string",
            "title": "Created By"
          },
          "sections": {
            "items": {
              "$ref": "#/components/schemas/BlueprintSectionResponse"
            },
            "type": "array",
            "title": "Sections",
            "default": []
          },
          "valid_from": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Valid From"
          },
          "valid_until": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Valid Until"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "blueprint_id",
          "competency_framework_version",
          "version",
          "status",
          "exam_duration_minutes",
          "total_items",
          "created_by"
        ],
        "title": "ExamBlueprintResponse"
      },
      "ExamBlueprintUpdate": {
        "properties": {
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          },
          "valid_until": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Valid Until"
          }
        },
        "type": "object",
        "title": "ExamBlueprintUpdate"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}